newBackend/requirements.txt
asgiref==3.3.4
Django==3.0
django-cors-headers==3.7.0
django-mssql-backend==2.8.1
djangorestframework==3.12.4
pyodbc==4.0.30
pytz==2021.1
sqlparse==0.4.1
xlwt==1.3.0
